ドキュメントが間違っているので修正
https://issuetracker.google.com/issues/358594663

+++ build-logic/src/main/kotlin/net/matsudamper/WatchFaceFormatGeneratorPlugin.kt
@@ -92,7 +92,7 @@ private fun SceneScope.CenterDigitalClock(
         ) {
             Font(
                 color = ContentColor.getColorSymbol(),
-                size = fontHeight.toFloat(),
+                size = fontHeight,
             )
         }
     }
