+++ build-logic/src/main/kotlin/net/matsudamper/dsl/scope/clock/TimeTextScope.kt
@@ -45,7 +45,7 @@ class TimeTextScope(
 
     fun Font(
         family: FontFamily = FontFamily.SyncToDevice,
-        size: Float,
+        size: Int,
         color: String,
         slant: FontSlant? = null,
         letterSpacing: Float = 0f,
